Change default dtype to int32 for Discrete

+++ gymnax/environments/spaces.py
@@ -26,7 +26,7 @@ class Discrete(Space):
     def __init__(
         self,
         num_categories: int,
-        dtype: jnp.dtype = jnp.int_
+        dtype: jnp.dtype = jnp.int32
     ):
         assert num_categories >= 0
         self.n = num_categories
